% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskFunction.R
\name{bind_multitree}
\alias{bind_multitree}
\title{Binding multiple trees}
\usage{
bind_multitree(...)
}
\arguments{
\item{...}{tree files}
}
\value{
a tree with multiple trees
}
\description{
This function binds multiple trees
}
