% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskFunction.R
\name{Dtime_Dist_CI_posterior_tree}
\alias{Dtime_Dist_CI_posterior_tree}
\title{Sample trees from BactDating posterior}
\usage{
Dtime_Dist_CI_posterior_tree(
  metadata,
  bd_res,
  initial_window = 5,
  Ntreestosample = 100,
  divtime_nboot = 20,
  upper_window = 50
)
}
\arguments{
\item{metadata}{metadata, should include GPSC, region, longitude, latitude, collection year}

\item{bd_res}{object from BactDating}

\item{initial_window}{window size of divergence time}

\item{Ntreestosample}{number of trees to sample from}

\item{divtime_nboot}{number of bootstrap sampling}

\item{upper_window}{maximum year of divergence}
}
\value{
a table of the relationship between pairwise divergence time and physical distance with confidence interval
}
\description{
This function is to sample trees from BactDating posterior to generate confidence interval
}
