% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskFunction.R
\name{gen_dist_mat}
\alias{gen_dist_mat}
\title{Generate distance matrix}
\usage{
gen_dist_mat(metadata, min_dist, max_dist)
}
\arguments{
\item{metadata}{metadata file with longitude and latitude information}

\item{min_dist}{minimum distance threshold}

\item{max_dist}{maximum distance threshold}
}
\value{
(1) a pairwise numeric distance matrix, and (2) a pairwise binary distance matrix
}
\description{
This function generates pairwise numeric distance matrix and binary distance matrix based on the given thresholds
}
