% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskFunction.R
\name{gen_divtime_mat}
\alias{gen_divtime_mat}
\title{Generate divergence time matrix}
\usage{
gen_divtime_mat(metadata, bactdate_tree, min_time, max_time)
}
\arguments{
\item{metadata}{metadata file with collection time information}

\item{bactdate_tree}{tree output generated by bactdacting}

\item{min_time}{minimum divergence time threshold}

\item{max_time}{maximum divergence time threshold}
}
\value{
(1) a pairwise numeric divergence time matrix, and (2) a pairwise binary divergence time matrix
}
\description{
This function generates pairwise numeric divergence time matrix and binary divergence time matrix based on the given thresholds
}
