% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskFunction.R
\name{Geo_Strain_Calc_RR}
\alias{Geo_Strain_Calc_RR}
\title{Calculate relative risk ratio}
\usage{
Geo_Strain_Calc_RR(
  x,
  des_geo_matrix,
  ref_geo_matrix,
  des_strain_matrix,
  ref_strain_matrix,
  colyear_matrix
)
}
\arguments{
\item{x}{subsampled isolates}

\item{des_geo_matrix}{designated geo_matrix generated by RR_matrix: des_location_mat or des_distance_mat}

\item{ref_geo_matrix}{reference geo_matrix generated by RR_matrix: ref_location_mat or ref_distance_mat}

\item{des_strain_matrix}{designated strain_mat generated by RR_matrix: des_lineage_mat or des_divtime_mat}

\item{ref_strain_matrix}{reference strain_mat generated by RR_matrix: ref_lineage_mat or ref_divtime_mat}

\item{colyear_matrix}{collection time matrix generated by RR_matrix: coltime_mat}
}
\value{
rr, relative risk ratio
}
\description{
This function calculates relative risk ratio with the input matrices generated by RR_matrix
}
