% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskFunction.R
\name{match_tree_meta}
\alias{match_tree_meta}
\title{Match Tree Label to Metadata Label}
\usage{
match_tree_meta(Metadata, Tree, drop_tip_name, drop_tip = TRUE)
}
\arguments{
\item{Metadata}{metadata file}

\item{Tree}{tree file generated from Gubbins}

\item{drop_tip_name}{the name of the tip to be dropped, usually the reference name}

\item{drop_tip}{TRUE to drop the tip of drop_tip_name}
}
\value{
a list with (1) a tree after dropping tip, and (2) a dataframe with matched tree and metadata label
}
\description{
This function generates pairwise matrices from metadata file.
}
