% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskFunction.R
\name{plt_gubbin_tree}
\alias{plt_gubbin_tree}
\title{Plot Gubbin Tree}
\usage{
plt_gubbin_tree(gubbin_tree)
}
\arguments{
\item{gubbin_tree}{tree file generated by Gubbins}
}
\value{
plot of Gubbins tree
}
\description{
This function generates Gubbins Tree plot.
}
