% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskFunction.R
\name{PltRR}
\alias{PltRR}
\title{Generate relative risk ratio plot}
\usage{
PltRR(rrData, RRplotTitle, x_title)
}
\arguments{
\item{rrData}{relative risk generated by RelativeRisk()}

\item{RRplotTitle}{name of the plot}

\item{x_title}{name of the x axis}
}
\value{
relative risk plot
}
\description{
This function plot relative risk ratio
}
