% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskFunction.R
\name{PltRR_rolling}
\alias{PltRR_rolling}
\title{Generate relative risk ratio plot with rolling distance}
\usage{
PltRR_rolling(rrData, RRplotTitle, xname)
}
\arguments{
\item{rrData}{relative risk generated by RelativeRisk()}

\item{RRplotTitle}{name of the plot}

\item{xname}{name of the x label}
}
\value{
relative risk plot with rolling distance
}
\description{
This function plot relative risk ratio with rolling distance
}
