% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskFunction.R
\name{PltRR_rolling_divtime}
\alias{PltRR_rolling_divtime}
\title{Generate relative risk ratio plot with rolling divergence time}
\usage{
PltRR_rolling_divtime(rrData, RRplotTitle)
}
\arguments{
\item{rrData}{relative risk generated by RelativeRisk()}

\item{RRplotTitle}{name of the plot}
}
\value{
relative risk plot with rolling distance
}
\description{
This function plot relative risk ratio with rolling divergence time
}
