% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskFunction.R
\name{RR_bootstrap}
\alias{RR_bootstrap}
\title{Calculate confidence interval of relative risk ratio}
\usage{
RR_bootstrap(
  metadataset,
  des_geo_matrix,
  ref_geo_matrix,
  des_strain_matrix,
  ref_strain_matrix,
  colyear_matrix,
  nboot = 20,
  subsample_min = 70
)
}
\arguments{
\item{metadataset}{metadata, should include GPSC, region, longitude, latitude, collection year}

\item{des_geo_matrix}{designated geo_matrix generated by RR_matrix: des_location_mat or des_distance_mat}

\item{ref_geo_matrix}{reference geo_matrix generated by RR_matrix: ref_location_mat or ref_distance_mat}

\item{des_strain_matrix}{designated strain_mat generated by RR_matrix: des_lineage_mat or des_divtime_mat}

\item{ref_strain_matrix}{reference strain_mat generated by RR_matrix: ref_lineage_mat or ref_divtime_mat}

\item{colyear_matrix}{collection time matrix generated by RR_matrix: coltime_mat}

\item{nboot}{number of bootstrapping}

\item{subsample_min}{number of subsample}
}
\value{
relative risk with confidence interval
}
\description{
This function calculates confidence interval of relative risk ratio by bootstrapping
}
