% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskFunction.R
\name{RR_matrix}
\alias{RR_matrix}
\title{Generate Pairwise Matrix for Relative Risk Calculation}
\usage{
RR_matrix(
  meta_table,
  bactdate_tree,
  coltime_range = 1,
  des_divtime_min = 0,
  des_divtime_max = 50,
  ref_divtime_min = 70,
  ref_divtime_max = 150,
  des_distance_min = 0,
  des_distance_max = 50,
  ref_distance_min = 150,
  ref_distance_max = 200
)
}
\arguments{
\item{meta_table}{metadata, should include GPSC, region, longitude, latitude, collection year}

\item{bactdate_tree}{tree file generated from bactdating}

\item{coltime_range}{given range of collection time}

\item{des_divtime_min}{given minimum designated divergence time}

\item{des_divtime_max}{given maximum designated divergence time}

\item{ref_divtime_min}{reference minimum divergence time}

\item{ref_divtime_max}{reference maximum divergence time}

\item{des_distance_min}{given minimum designated distance}

\item{des_distance_max}{given maximum designated distance}

\item{ref_distance_min}{reference minimum distance}

\item{ref_distance_max}{reference maximum distance}
}
\value{
pairwise matrices, including (1) pairwise designated lineage matrix, (2) pairwise reference lineage matrix, (3) pairwise designated location matrix, (4) pairwise reference location matrix, (5) pairwise designated binary distance matrix, (6) pairwise reference binary distance matrix, (7) pairwise designated binary divergence time matrix, (8) pairwise reference binary divergence time matrix, (9) pairwise numeric distance matrix, (10) pairwise divergence time matrix, (11) pairwise collection time matrix
}
\description{
This function generates pairwise matrices from metadata file and bactdating tree
}
