% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskFunction.R
\name{RR_matrix_notree}
\alias{RR_matrix_notree}
\title{Generate Pairwise Matrix for Relative Risk Calculation without tree}
\usage{
RR_matrix_notree(
  meta_table,
  coltime_range = 1,
  des_distance_min = 0,
  des_distance_max = 50,
  ref_distance_min = 200,
  ref_distance_max = 300
)
}
\arguments{
\item{meta_table}{metadata, should include GPSC, region, longitude, latitude, collection year}

\item{coltime_range}{given range of collection time}

\item{des_distance_min}{given minimum designated distance}

\item{des_distance_max}{given maximum designated distance}

\item{ref_distance_min}{reference minimum distance}

\item{ref_distance_max}{reference maximum distance}
}
\value{
pairwise matrices, including (1) pairwise designated lineage matrix, (2) pairwise reference lineage matrix, (3) pairwise designated location matrix, (4) pairwise reference location matrix, (5) pairwise designated binary distance matrix, (6) pairwise reference binary distance matrix, (7) pairwise numeric distance matrix, (8) pairwise divergence time matrix, (9) pairwise collection time matrix
}
\description{
This function generates pairwise matrices from metadata file and bactdating tree
}
