% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskFunction.R
\name{to.upper}
\alias{to.upper}
\title{Extract upper diagnal matrix}
\usage{
to.upper(X)
}
\arguments{
\item{X}{input matrix}
}
\value{
upper diagnal matrix
}
\description{
This function extracts upper diagnal matrix
}
